+++ kads/__main__.py
@@ -2,5 +2,5 @@ import sys
 
 from kads.cli import main
 
-if __name__ == "__main__":
-    sys.exit(main())
+if __name__ == "__main__":  # pragma: no cover
+    sys.exit(main())  # pragma: no cover